--- final_project/Being.cpp
@@ -1,2 +1,0 @@
-#include "Being.h"
-
