--- final_project/Being.h
@@ -1,15 +1,0 @@
-#pragma once
-
-#include <string>
-
-using namespace std;
-
-class Being
-{
-public:
-	string name;
-	int hp;
-    int maxhp;
-
-};
-
