--- final_project/BeingData.cpp
@@ -1,1 +1,0 @@
-#include "BeingData.h"
