--- final_project/BeingData.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <string>
-
-using namespace std;
-
-class BeingData
-{
-public:
-	int hp;
-	string	name;
-};
-
