--- final_project/Character.cpp
@@ -1,13 +1,0 @@
-#include <iostream>
-#include "Character.h"
-
-Character::Character(CharcterData& cdata, int skillsNum[])
-{
-	this->name = cdata.name;
-	this->hp = cdata.hp;
-
-	for (int i = 0; i < cdata.maxSkill; i++) {
-		this->skills.push_back(cdata.skills[skillsNum[i]]);
-	}
-
-}
