+++ final_project/Header/Being.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <string>
+
+using namespace std;
+
+class Being
+{
+public:
+	string name;
+	int hp;
+    int maxhp;
+
+};
+
