+++ final_project/Header/BeingData.h
@@ -1,0 +1,12 @@
+#pragma once
+#include <string>
+
+using namespace std;
+
+class BeingData
+{
+public:
+	int hp;
+	string	name;
+};
+
