+++ final_project/Header/Monster.h
@@ -1,0 +1,20 @@
+#pragma once
+#include <vector>
+#include "Postition.h"
+#include "MonsterSkill.h"
+#include "Being.h"
+
+using namespace std;
+
+class Monster :public Being
+{
+
+public:
+	int atk, range;
+	vector<MonsterSkill> skills;
+	Postition pos;
+
+	Monster();
+	Monster(int hp, int atk, int range);
+};
+
