+++ final_project/Header/Skill.h
@@ -1,0 +1,11 @@
+#pragma once
+
+using namespace std;
+
+class Skill
+{
+public:
+	int number;
+	int sp;
+};
+
