--- final_project/MonsterData.cpp
@@ -1,6 +1,0 @@
-#include <fstream>
-#include <iostream>
-
-#include "MonsterData.h"
-
-using namespace std;
