--- final_project/Postition.cpp
@@ -1,13 +1,0 @@
-#include "Postition.h"
-
-Postition::Postition()
-{
-	this->x = 0;
-	this->y = 0;
-}
-
-Postition::Postition(int x, int y)
-{
-	this->x = x;
-	this->y = y;
-}
