--- final_project/Skill.cpp
@@ -1,1 +1,0 @@
-#include "Skill.h"
