--- final_project/Skill.h
@@ -1,11 +1,0 @@
-#pragma once
-
-using namespace std;
-
-class Skill
-{
-public:
-	int number;
-	int sp;
-};
-
