+++ final_project/define/Being.cpp
@@ -1,0 +1,2 @@
+#include "Being.h"
+
