+++ final_project/define/BeingData.cpp
@@ -1,0 +1,1 @@
+#include "BeingData.h"
