+++ final_project/define/MonsterData.cpp
@@ -1,0 +1,6 @@
+#include <fstream>
+#include <iostream>
+
+#include "MonsterData.h"
+
+using namespace std;
