+++ final_project/input/charcter.txt
@@ -1,0 +1,19 @@
+2
+brute 14 6
+8
+0 05 attack 4 - shield 2
+1 14 attack 3 range 3 - move 3
+2 20 attack 3 - move 4 heal 1
+3 28 attack 2 shield 1 - move 3
+4 37 attack 2 attack 2 - move 1 shield 1
+5 49 shield 2 - attack 3 range 2
+6 56 heal 3 - move 3
+7 88 attack 3 - move 3
+scoundrel 8 5
+6
+0 04 attack 5 - move 3
+1 08 attack 4 - move 3
+2 27 attack 3 - move 3
+3 38 attack 2 attack 2 attack 2 - move 4
+4 44 attack 7 - move 2 attack 3 move 2
+5 68 attack 4 range 4 - heal 3
