+++ final_project/input/map.txt
@@ -1,0 +1,21 @@
+12 9
+000000000
+000111000
+000111000
+000111000
+000111000
+000111000
+000030000
+011111110
+011111110
+011111110
+011111110
+000000000
+3 10 4 10 5 10 4 9
+6
+archer 3 1 0 1 1
+archer 4 1 2 2 2
+archer 5 1 0 1 1
+guard 3 7 0 1 1
+guard 4 7 1 1 2
+guard 5 7 0 1 1
